move much non-tauri code to separate module
