+++ .vscode/settings.json
@@ -60,7 +60,7 @@
   ],
   "editor.inlayHints.enabled": "off",
   "editor.tabSize": 2,
-  "rust-analyzer.cargo.features": "all",
+  "rust-analyzer.cargo.features": [],
   "css.completion.completePropertyWithSemicolon": false,
   "rust-analyzer.cargo.targetDir": true,
   "python.terminal.activateEnvironment": false
