+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "3"
-members = ["src-tauri", "pkm_rs", "pkm_rs_derive", "pkm_rs_resources", "pkm_rs_types"]
+members = ["src-tauri", "pkm_rs", "pkm_rs_derive", "pkm_rs_resources", "pkm_rs_types", "openhome_core"]
 
 [profile.release]
 panic = "abort"
