+++ openhome_core/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "openhome_core"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+base64 = "0.22.1"
+pkm_rs = {path = "../pkm_rs"}
+pkm_rs_derive = {path = "../pkm_rs_derive"}
+semver = "1.0.28"
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.150"
+specta = { version = "=2.0.0-rc.22", features = ["chrono", "serde_json", "uuid" ], optional = true }
+strum = "0.28.0"
+tokio = { version = "1.52.3" }
+tracing = "0.1.44"
+uuid = { version = "1.24.0", features = ["v4"] }
+
+[lib]
+name = "openhome_core"
+path = "src/lib.rs"
+
+[features]
+desktop = ["specta"]
