+++ openhome_core/src/error.rs
@@ -45,7 +45,6 @@ pub enum Error {
         context: String,
         source: Option<SourceError>,
     },
-    Tauri(tauri::Error),
 }
 
 impl Error {
@@ -189,9 +188,6 @@ impl Display for Error {
                 Some(source) => format!("{context}: {source}"),
                 None => context.clone(),
             },
-            Self::Tauri(source) => {
-                format!("Tauri error: ({source})")
-            }
             Self::UnexpectedCondition { context, source } => {
                 let source_str = source
                     .as_ref()
@@ -233,12 +229,6 @@ impl std::error::Error for Error {
 impl<T> From<std::sync::PoisonError<T>> for Error {
     fn from(_: std::sync::PoisonError<T>) -> Self {
         Error::MutexFailure
-    }
-}
-
-impl From<tauri::Error> for Error {
-    fn from(source: tauri::Error) -> Self {
-        Self::Tauri(source)
     }
 }
