+++ src-tauri/src/commands.rs
@@ -1,9 +1,12 @@
-use crate::data_controller::{DataController, DataDir};
-use crate::error::{Error, Result};
+use crate::data_controller::ToDataController;
 use crate::plugin::{self, PluginMetadata, PluginMetadataWithIcon, list_downloaded_plugins};
 use crate::state::{AppState, AppStateInner};
 use crate::util::ImageResponse;
-use crate::{menu, saves, util};
+use crate::{menu, util};
+use openhome_core::data_controller::{DataController, DataDir};
+use openhome_core::error::{Error, Result};
+use openhome_core::pkm_storage::StoredBankData;
+use openhome_core::saves::{self, SaveFileSearch};
 use serde_json::Value;
 use std::fs;
 use std::path::{Path, PathBuf};
@@ -62,7 +65,9 @@ fn write_storage_file_text(
     relative_path: PathBuf,
     text: String,
 ) -> Result<()> {
-    let full_path = app_handle.absolute_path(DataDir::Storage, relative_path)?;
+    let full_path = app_handle
+        .controller()
+        .absolute_path(DataDir::Storage, relative_path)?;
     util::write_file_contents(full_path, text)
 }
 
@@ -71,7 +76,9 @@ pub fn get_storage_file_json(
     app_handle: tauri::AppHandle,
     relative_path: PathBuf,
 ) -> Result<Value> {
-    app_handle.read_file_json(DataDir::Storage, &relative_path)
+    app_handle
+        .controller()
+        .read_file_json(DataDir::Storage, &relative_path)
 }
 
 #[tauri::command]
@@ -116,9 +123,9 @@ pub fn set_app_theme(
 #[tauri::command]
 #[specta::specta]
 pub fn validate_recent_saves(
-    mut app_handle: tauri::AppHandle,
+    app_handle: tauri::AppHandle,
 ) -> CommandResult<Vec<(String, saves::SaveRef)>> {
-    Ok(saves::get_recent_saves(&mut app_handle)?)
+    Ok(saves::get_recent_saves(&app_handle.controller())?)
 }
 
 #[tauri::command]
@@ -160,20 +167,23 @@ pub async fn download_plugin(
 pub fn list_installed_plugins(
     app_handle: tauri::AppHandle,
 ) -> CommandResult<Vec<PluginMetadataWithIcon>> {
-    Ok(list_downloaded_plugins(&app_handle)?)
+    Ok(list_downloaded_plugins(&app_handle.controller())?)
 }
 
 #[tauri::command]
 #[specta::specta]
 pub fn load_plugin_code(app_handle: tauri::AppHandle, plugin_id: String) -> CommandResult<String> {
     let relative_path = &PathBuf::from(plugin_id).join("dist").join("index.js");
-    Ok(app_handle.read_file_text(DataDir::Plugins, relative_path)?)
+    Ok(app_handle
+        .controller()
+        .read_file_text(DataDir::Plugins, relative_path)?)
 }
 
 #[tauri::command]
 #[specta::specta]
 pub fn delete_plugin(app_handle: tauri::AppHandle, plugin_id: String) -> CommandResult<()> {
     let plugin_dir = app_handle
+        .controller()
         .absolute_dir_path(DataDir::Plugins)?
         .join(&plugin_id);
     Ok(util::delete_directory(&plugin_dir)?)
@@ -183,4 +193,75 @@ pub fn delete_plugin(app_handle: tauri::AppHandle, plugin_id: String) -> Command
 #[specta::specta]
 pub fn handle_windows_accelerator(app_handle: tauri::AppHandle, menu_event_id: String) {
     menu::handle_menu_event_id(&app_handle, menu_event_id.as_ref());
+}
+
+#[tauri::command]
+#[specta::specta]
+pub async fn find_suggested_saves(
+    app_handle: tauri::AppHandle,
+    save_folders: Vec<&str>,
+) -> CommandResult<saves::PossibleSaves> {
+    let mut possible_saves = saves::PossibleSaves {
+        citra: Vec::new(),
+        desmume: Vec::new(),
+        open_emu: Vec::new(),
+    };
+
+    let citra_dir_r = app_handle
+        .path()
+        .home_dir()
+        .map(|home| home.join(".local/share/citra-emu/sdmc/Nintendo 3DS"));
+
+    if let Ok(citra_dir) = citra_dir_r
+        && citra_dir.exists()
+    {
+        possible_saves
+            .citra
+            .extend(saves::Citra::recursively_find_saves(&citra_dir)?);
+    }
+
+    // Iterate over user-provided save folders
+    for folder_str in save_folders {
+        let folder_path = PathBuf::from(folder_str);
+        if folder_path.exists() {
+            tracing::info!("checking saves in folder {folder_str}");
+            let result = tokio::task::spawn_blocking(move || {
+                openhome_core::saves::get_possible_saves(&folder_path).map_err(|e| e.to_string())
+            })
+            .await
+            .map_err(|e| {
+                Error::other_with_source("tokio task failed in find_suggested_saves", e)
+            })?;
+            tracing::info!("finished checking");
+
+            match result {
+                Ok(newly_found) => possible_saves.add_all(newly_found),
+                Err(e) => {
+                    tracing::error!("failed to check saves in folder {folder_str}: {e}");
+                    continue;
+                }
+            };
+        } else {
+            return Err(Error::file_missing(&folder_path).into());
+        }
+    }
+
+    Ok(possible_saves)
+}
+
+#[tauri::command]
+#[specta::specta]
+pub fn load_banks(app_handle: tauri::AppHandle) -> CommandResult<StoredBankData> {
+    Ok(openhome_core::pkm_storage::load_banks(
+        &app_handle.controller(),
+    )?)
+}
+
+#[tauri::command]
+#[specta::specta]
+pub fn write_banks(app_handle: tauri::AppHandle, bank_data: StoredBankData) -> CommandResult<()> {
+    Ok(openhome_core::pkm_storage::write_banks(
+        &app_handle.controller(),
+        bank_data,
+    )?)
 }
