+++ src-tauri/src/deprecated/pre_1_5_0.rs
@@ -1,7 +1,6 @@
-use crate::pkm_storage;
-use std::collections::HashMap;
-
+use openhome_core::pkm_storage;
 use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
 use uuid::Uuid;
 
 pub const BOXDATA_FILE: &str = "box-data.json";
