+++ src-tauri/src/synced_state/convert_strategies.rs
@@ -1,39 +1,9 @@
-use std::collections::HashMap;
-
-use crate::{
-    commands::{CommandError, CommandResult},
-    data_controller::{DataController, DataDir},
-    error::Result,
-};
-use pkm_rs::convert_strategy::ConvertStrategy;
+use crate::commands::{CommandError, CommandResult};
+use openhome_core::convert_strategies::{ConvertStrategies, NamedStrategy};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::synced_state;
-
-#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
-pub struct NamedStrategy {
-    name: String,
-    strategy: ConvertStrategy,
-}
-
-impl NamedStrategy {
-    #[cfg(test)]
-    pub fn inner(&self) -> &ConvertStrategy {
-        &self.strategy
-    }
-}
-
-type StrategiesById = HashMap<Uuid, NamedStrategy>;
-
-#[derive(Debug, Default, Clone, Serialize, Deserialize, specta::Type)]
-pub struct ConvertStrategies {
-    strategies_by_id: StrategiesById,
-    default_strategy_id: Option<Uuid>,
-}
-
-pub const DATA_DIR: DataDir = DataDir::Storage;
-pub const JSON_FILENAME: &str = "convert_strategies.json";
 
 impl synced_state::SyncedState for ConvertStrategies {
     type Action = ConvertStrategyEntries;
@@ -41,28 +11,13 @@ impl synced_state::SyncedState for ConvertStrategies {
 
     fn update(&mut self, action: Self::Action) {
         for (key, value) in action.ids_and_strategies {
-            self.strategies_by_id.insert(key, value);
+            self.insert(key, value);
         }
-        self.default_strategy_id = Some(action.default_strategy_id);
+        self.set_default(action.default_strategy_id);
     }
 
     fn to_command_response(&self) -> impl Clone + Serialize + tauri::ipc::IpcResponse {
         self
-    }
-}
-
-impl ConvertStrategies {
-    pub fn load_from_storage(data_controller: &mut impl DataController) -> Result<Self> {
-        data_controller.read_or_create_default_json_file(DATA_DIR, JSON_FILENAME)
-    }
-
-    pub fn write_to_files(&self, data_controller: &mut impl DataController) -> Result<()> {
-        data_controller.write_file_json(DATA_DIR, JSON_FILENAME, self)
-    }
-
-    #[cfg(test)]
-    pub fn get(&self, key: Uuid) -> Option<&NamedStrategy> {
-        self.strategies_by_id.get(&key)
     }
 }
 
@@ -80,8 +35,8 @@ pub fn get_convert_strategies(
     synced_state
         .get_convert_strategies()
         .map(|strategies| ConvertStrategyEntries {
-            ids_and_strategies: strategies.strategies_by_id.into_iter().collect(),
-            default_strategy_id: strategies.default_strategy_id.unwrap_or_default(),
+            ids_and_strategies: strategies.all(),
+            default_strategy_id: strategies.default_id().unwrap_or_default(),
         })
         .map_err(CommandError::from)
 }
