+++ src-tauri/src/synced_state/ohpkm_store.rs
@@ -1,126 +1,12 @@
 use crate::commands::CommandResult;
-use crate::data_controller::{DataController, DataDir, MONS_V2_DIR};
-use crate::error::{Error, Result};
+use crate::data_controller::TauriDataController;
 use crate::synced_state;
-use crate::util;
-use base64::prelude::*;
-use pkm_rs::ohpkm::OhpkmV2;
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-use std::num::NonZeroU64;
-use std::path::{Path, PathBuf};
-use std::time::{Duration, UNIX_EPOCH};
+use openhome_core::Error;
+use openhome_core::data_controller::{DataController, DataDir, MONS_V2_DIR};
+use openhome_core::ohpkm_store::OhpkmBytesStore;
+use serde::Serialize;
+use std::path::Path;
 use std::{collections::HashMap, fs};
-use tracing::warn;
-
-#[derive(Default, Serialize, Deserialize, Clone, specta::Type)]
-pub struct OhpkmBytesStore(HashMap<String, Vec<u8>>);
-
-impl OhpkmBytesStore {
-    fn load_from_directory(path: &Path) -> Result<Self> {
-        let mon_files = fs::read_dir(path).map_err(|e| Error::file_access(&path, e))?;
-
-        let mut map = HashMap::new();
-        for dir_entry in mon_files.flatten() {
-            let path = dir_entry.path();
-            if !path
-                .extension()
-                .is_some_and(|ext| ext.eq_ignore_ascii_case("ohpkm"))
-            {
-                continue;
-            }
-
-            if let Ok(mon_bytes) = util::read_file_bytes(path)
-                && let Ok(mut mon) = OhpkmV2::from_bytes(&mon_bytes)
-            {
-                let file_created_seconds = dir_entry
-                    .metadata()
-                    .ok()
-                    .and_then(|m| m.created().or(m.modified()).ok())
-                    .and_then(|t| t.duration_since(UNIX_EPOCH).ok())
-                    .as_ref()
-                    .map(Duration::as_secs)
-                    .and_then(NonZeroU64::new);
-
-                mon.set_started_tracking_if_missing(file_created_seconds);
-                map.insert(mon.openhome_id(), mon.to_bytes());
-            }
-        }
-
-        let mut store = Self(map);
-
-        store.fix_errors();
-
-        Ok(store)
-    }
-
-    fn fix_errors(&mut self) {
-        for (identifier, bytes) in self.0.iter_mut() {
-            if let Ok(mut mon) = OhpkmV2::from_bytes(bytes) {
-                let errors = mon.fix_errors();
-                if !errors.is_empty() {
-                    let errors_fixed_msgs: Vec<String> =
-                        errors.into_iter().map(|e| e.to_string()).collect();
-                    let errors_fixed_serialized = json!({"errors_fixed": errors_fixed_msgs});
-                    warn!(event = "ohpkm_errors_fixed", context = %errors_fixed_serialized, ohpkm_id = mon.openhome_id(), "Fixed Ohpkm {identifier} with nickname {}", mon.get_nickname());
-                    *bytes = mon.to_bytes();
-                }
-            }
-        }
-    }
-
-    fn write_to_directory(data: &Self, path: &Path) -> Result<()> {
-        let mut errors: Vec<(PathBuf, Box<dyn std::error::Error>)> = Vec::new();
-        data.0.iter().for_each(|(identifier, bytes)| {
-            let filename = format!("{identifier}.ohpkm");
-            let file_path = path.join(filename);
-            if let Err(err) = fs::write(&file_path, bytes) {
-                errors.push((file_path, Box::new(err)));
-            };
-        });
-
-        if !errors.is_empty() {
-            Err(Error::FileWrites(errors))
-        } else {
-            Ok(())
-        }
-    }
-
-    pub fn load_from_mons_v2(data_controller: &impl DataController) -> Result<Self> {
-        let mons_v2_dir = data_controller.absolute_path(DataDir::Storage, MONS_V2_DIR)?;
-        Self::load_from_directory(&mons_v2_dir)
-    }
-
-    pub fn write_to_mons_v2(&self, data_controller: &impl DataController) -> Result<()> {
-        let mons_v2_dir = data_controller.absolute_path(DataDir::Storage, MONS_V2_DIR)?;
-        Self::write_to_directory(self, &mons_v2_dir)
-    }
-
-    pub fn to_b64_map(&self) -> HashMap<String, String> {
-        let mut output: HashMap<String, String> = HashMap::new();
-        for (k, v) in self.0.clone() {
-            output.insert(k, BASE64_STANDARD.encode(v));
-        }
-
-        output
-    }
-
-    pub fn to_b64_entries(&self) -> Vec<(String, String)> {
-        self.0
-            .clone()
-            .into_iter()
-            .map(|(k, v)| (k, BASE64_STANDARD.encode(v)))
-            .collect()
-    }
-
-    pub fn includes(&self, identifier: &str) -> bool {
-        self.0.contains_key(identifier)
-    }
-
-    pub fn remove(&mut self, identifier: &str) -> bool {
-        self.0.remove(identifier).is_some()
-    }
-}
 
 impl synced_state::SyncedState for OhpkmBytesStore {
     type Action = Self;
@@ -131,8 +17,8 @@ impl synced_state::SyncedState for OhpkmBytesStore {
     }
 
     fn update(&mut self, other: Self) {
-        other.0.into_iter().for_each(|(k, v)| {
-            self.0.insert(k, v);
+        other.all_entries().for_each(|(k, v)| {
+            self.insert(k, v);
         });
     }
 }
@@ -180,10 +66,11 @@ pub fn permanently_delete_ohpkms(
         })?;
     let mut results = HashMap::new();
 
+    let controller = TauriDataController::new(&app_handle);
     // then delete from the disk
     for identifier in openhome_ids {
         let relative_path = Path::new(MONS_V2_DIR).join(format!("{identifier}.ohpkm"));
-        match app_handle.absolute_path(DataDir::Storage, &relative_path) {
+        match controller.absolute_path(DataDir::Storage, &relative_path) {
             Ok(full_path) => {
                 let deletion_result = fs::remove_file(full_path)
                     .map_err(|e| Error::file_access(&relative_path, e).to_string());
