+++ src-tauri/src/util.rs
@@ -1,14 +1,12 @@
 use base64::Engine;
 use base64::engine::general_purpose;
-use std::collections::HashSet;
 use std::fs;
-use std::hash::{Hash, Hasher};
 use std::io::Cursor;
 use std::path::Path;
 use std::process::Command;
 use zip::ZipArchive;
 
-use crate::error::{Error, Result};
+use openhome_core::{Error, Result};
 
 #[cfg(target_os = "linux")]
 use dialog::DialogBox;
@@ -19,54 +17,6 @@ use tauri_plugin_dialog::{DialogExt, MessageDialogKind};
 pub struct ImageResponse {
     pub base64: String,
     pub extension: String,
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug, specta::Type)]
-pub struct PathData {
-    pub raw: String,
-    pub name: String,
-    pub dir: String,
-    pub ext: String,
-    pub separator: String,
-}
-
-impl PartialEq for PathData {
-    fn eq(&self, other: &Self) -> bool {
-        self.raw == other.raw
-    }
-}
-
-impl Eq for PathData {}
-impl Hash for PathData {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.raw.hash(state);
-    }
-}
-
-pub fn parse_path_data(path: &Path) -> PathData {
-    let raw = path.to_string_lossy().to_string();
-    let name = path
-        .file_stem()
-        .unwrap_or_default()
-        .to_string_lossy()
-        .into();
-    let ext = path
-        .extension()
-        .unwrap_or_default()
-        .to_string_lossy()
-        .into();
-    let dir = path
-        .parent()
-        .map(|p| p.to_string_lossy().to_string())
-        .unwrap_or_default();
-    let separator = std::path::MAIN_SEPARATOR.to_string();
-    PathData {
-        raw,
-        name,
-        dir,
-        ext,
-        separator,
-    }
 }
 
 pub fn write_file_contents<P, C>(path: P, contents: C) -> Result<()>
@@ -89,11 +39,6 @@ where
     P: AsRef<Path>,
 {
     fs::read(&path).map_err(|err| Error::file_access(&path, err))
-}
-
-pub fn dedupe_paths(paths: &mut Vec<PathData>) {
-    let set: HashSet<PathData> = paths.drain(..).collect();
-    *paths = set.into_iter().collect()
 }
 
 pub async fn download_text_file(url: &str) -> Result<String> {
